+++ extentreports-cucumberjson-plugin/src/test/java/tech/grasshopper/test/heirarchy/ExtentTestHeirarchyTest.java
@@ -66,11 +66,8 @@ public class ExtentTestHeirarchyTest {
 		String startTime = "2020-06-20T12:30:15.654Z";
 		String endTime = "2020-06-20T12:30:16.123Z";
 
-		Feature feature = (new FeatureBuilder())
-				.setUri(uri).setName(name).setDescription(description)
-				.setStartTime(startTime).setEndTime(endTime)
-				.addFeatureTag(tagName1).addFeatureTag(tagName2)
-				.build();
+		Feature feature = (new FeatureBuilder()).setUri(uri).setName(name).setDescription(description)
+				.setStartTime(startTime).setEndTime(endTime).addFeatureTag(tagName1).addFeatureTag(tagName2).build();
 
 		ExtentTest featureExtentTest = extentTestHeirarchy.createFeatureExtentTest(feature);
 		com.aventstack.extentreports.model.Test model = featureExtentTest.getModel();
@@ -78,7 +75,7 @@ public class ExtentTestHeirarchyTest {
 		assertEquals("Test name is not correct.", name, model.getName());
 		assertEquals("Test description is not correct.", description, model.getDescription());
 		assertEquals("Test categories are not correct.", Arrays.asList(tagName1, tagName2),
-				model.getCategoryContext().getAll().stream().map(t -> t.getName()).collect(Collectors.toList()));
+				model.getCategorySet().stream().map(t -> t.getName()).collect(Collectors.toList()));
 		assertEquals("Test start time is not correct.", DateConverter.parseToDate(startTime), model.getStartTime());
 		assertEquals("Test end time is not correct.", DateConverter.parseToDate(endTime), model.getEndTime());
 	}
@@ -89,36 +86,28 @@ public class ExtentTestHeirarchyTest {
 		String name = "Feature Name";
 		String startTime = "2020-06-20T12:30:15.654Z";
 		String endTime = "2020-06-20T12:30:16.123Z";
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri(uri).setName(name)
-				.setStartTime(startTime).setEndTime(endTime)
+
+		Feature feature = (new FeatureBuilder()).setUri(uri).setName(name).setStartTime(startTime).setEndTime(endTime)
 				.build();
 
 		ExtentTest featureExtentTest = extentTestHeirarchy.createFeatureExtentTest(feature);
 		com.aventstack.extentreports.model.Test model = featureExtentTest.getModel();
 
 		assertEquals("Test name is not correct.", name, model.getName());
-		assertTrue("Test categories are not empty.", model.getCategoryContext().getAll().isEmpty());
+		assertTrue("Test categories are not empty.", model.getCategorySet().isEmpty());
 		assertEquals("Test start time is not correct.", DateConverter.parseToDate(startTime), model.getStartTime());
 		assertEquals("Test end time is not correct.", DateConverter.parseToDate(endTime), model.getEndTime());
 	}
 
 	@Test
 	public void testRepeatFeatureCreation() {
-		String name = "Feature Name";		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature")
-				.setName(name)
-				.setStartTime("2020-06-20T12:30:15.654Z")
-				.setEndTime("2020-06-20T12:30:16.123Z")
-				.build();
+		String name = "Feature Name";
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName(name)
+				.setStartTime("2020-06-20T12:30:15.654Z").setEndTime("2020-06-20T12:30:16.123Z").build();
 
 		ExtentTest featureExtentTest = extentTestHeirarchy.createFeatureExtentTest(feature);
 
-		Feature featureRepeat = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature")
-				.setName(name)
+		Feature featureRepeat = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName(name)
 				.build();
 
 		ExtentTest featureExtentTestRepeat = extentTestHeirarchy.createFeatureExtentTest(featureRepeat);
@@ -137,46 +126,40 @@ public class ExtentTestHeirarchyTest {
 		String description = "Scenario Description";
 		String tagName1 = "tag1";
 		String tagName2 = "tag2";
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario")
-				.setScenarioName(name).setScenarioDescription(description)
-				.addScenarioTag(tagName1).addScenarioTag(tagName2)
-				.build();
 
-		ExtentTest scenarioExtentTest = extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest, feature.getElements().get(0));
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario").setScenarioName(name).setScenarioDescription(description)
+				.addScenarioTag(tagName1).addScenarioTag(tagName2).build();
+
+		ExtentTest scenarioExtentTest = extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest,
+				feature.getElements().get(0));
 		com.aventstack.extentreports.model.Test model = scenarioExtentTest.getModel();
 
 		assertEquals("Test name is not correct.", name, model.getName());
 		assertEquals("Test description is not correct.", description, model.getDescription());
 		assertEquals("Test categories are not correct.", Arrays.asList(tagName1, tagName2),
-				model.getCategoryContext().getAll().stream().map(t -> t.getName()).collect(Collectors.toList()));
+				model.getCategorySet().stream().map(t -> t.getName()).collect(Collectors.toList()));
 
 		List<com.aventstack.extentreports.model.Test> featureTestChilds = parentFeatureExtentTest.getModel()
-				.getNodeContext().getAll();
+				.getChildren();
 		assertEquals("Number of child nodes of Feature Extent Test should be 1", 1, featureTestChilds.size());
 		assertEquals("Scenario Extent Test should be child of Feature Extent Test.", model, featureTestChilds.get(0));
 	}
-	
+
 	@Test
 	public void testMultipleNewScenarioCreation() {
 		ExtentTest parentFeatureExtentTest = extent.createTest(com.aventstack.extentreports.gherkin.model.Feature.class,
 				"Parent Feature Test", "");
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario")
-				.setScenarioName("Scenario Name")
-				.addScenario().setScenarioKeyword("scenario")
-				.setScenarioName("Scenario Name Repeat")
-				.build();
+
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario").setScenarioName("Scenario Name").addScenario()
+				.setScenarioKeyword("scenario").setScenarioName("Scenario Name Repeat").build();
 
 		extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest, feature.getElements().get(0));
 		extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest, feature.getElements().get(1));
 
 		List<com.aventstack.extentreports.model.Test> featureTestChilds = parentFeatureExtentTest.getModel()
-				.getNodeContext().getAll();
+				.getChildren();
 		assertEquals("Number of child nodes of Feature Extent Test should be 2", 2, featureTestChilds.size());
 	}
 
@@ -187,14 +170,11 @@ public class ExtentTestHeirarchyTest {
 
 		String name = "Scenario Name";
 		String description = "Scenario Description";
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario outline")
-				.setScenarioName(name).setScenarioDescription(description)
-				.setScenarioId("Feature-Name;Scenario-Outline-Name;Examples-Name;2")
-				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":5:6:7")
-				.build();
+
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario outline").setScenarioName(name)
+				.setScenarioDescription(description).setScenarioId("Feature-Name;Scenario-Outline-Name;Examples-Name;2")
+				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":5:6:7").build();
 
 		ExtentTest scenarioExtentTest = extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest,
 				feature.getElements().get(0));
@@ -202,18 +182,17 @@ public class ExtentTestHeirarchyTest {
 
 		assertEquals("Test name is not correct.", name, model.getName());
 		assertEquals("Test description is not correct.", description, model.getDescription());
-		assertTrue("Test categories are not empty.", model.getCategoryContext().getAll().isEmpty());
+		assertTrue("Test categories are not empty.", model.getCategorySet().isEmpty());
 
 		List<com.aventstack.extentreports.model.Test> scenarioOutlineTest = parentFeatureExtentTest.getModel()
-				.getNodeContext().getAll();
+				.getChildren();
 		assertEquals("Number of child nodes of Feature Extent Test should be 1", 1, scenarioOutlineTest.size());
 		assertEquals("Scenario Outline Test name is not correct.", "Scenario Outline Name",
 				scenarioOutlineTest.get(0).getName());
 		assertEquals("Scenario Outline Test description is not correct.", description,
 				scenarioOutlineTest.get(0).getDescription());
 
-		List<com.aventstack.extentreports.model.Test> scenarioTest = scenarioOutlineTest.get(0).getNodeContext()
-				.getAll();
+		List<com.aventstack.extentreports.model.Test> scenarioTest = scenarioOutlineTest.get(0).getChildren();
 		assertEquals("Number of child nodes of Scenario Outline Test should be 1", 1, scenarioTest.size());
 		assertEquals("Scenario Outline Extent Test should be child of Scenario Extent Test.", model,
 				scenarioTest.get(0));
@@ -223,27 +202,22 @@ public class ExtentTestHeirarchyTest {
 	public void testMultipleScenarioForSingleScenarioOutlineCreation() {
 		ExtentTest parentFeatureExtentTest = extent.createTest(com.aventstack.extentreports.gherkin.model.Feature.class,
 				"Parent Feature Test", "");
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario outline")
-				.setScenarioName("Scenario Outline Name")
+
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario outline").setScenarioName("Scenario Outline Name")
 				.setScenarioId("Feature-Name;Scenario-Outline-Name;Examples-Name;2")
-				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":5:6:7")
-				.addScenario().setScenarioKeyword("scenario outline")
-				.setScenarioName("Scenario Outline Name Repeat")
+				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":5:6:7").addScenario()
+				.setScenarioKeyword("scenario outline").setScenarioName("Scenario Outline Name Repeat")
 				.setScenarioId("Feature-Name;Scenario-Outline-Name;Examples-Name;3")
-				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":5:6:7")
-				.build();
+				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":5:6:7").build();
 
 		extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest, feature.getElements().get(0));
 		extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest, feature.getElements().get(1));
 
 		List<com.aventstack.extentreports.model.Test> scenarioOutlineTest = parentFeatureExtentTest.getModel()
-				.getNodeContext().getAll();
+				.getChildren();
 		assertEquals("Number of child nodes of Feature Extent Test should be 1", 1, scenarioOutlineTest.size());
-		List<com.aventstack.extentreports.model.Test> scenarioTest = scenarioOutlineTest.get(0).getNodeContext()
-				.getAll();
+		List<com.aventstack.extentreports.model.Test> scenarioTest = scenarioOutlineTest.get(0).getChildren();
 		assertEquals("Number of child nodes of Scenario Outline Test should be 2", 2, scenarioTest.size());
 	}
 
@@ -251,32 +225,26 @@ public class ExtentTestHeirarchyTest {
 	public void testMultipleScenarioOutlineCreation() {
 		ExtentTest parentFeatureExtentTest = extent.createTest(com.aventstack.extentreports.gherkin.model.Feature.class,
 				"Parent Feature Test", "");
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario outline")
-				.setScenarioName("Scenario Outline Name First")
+
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario outline").setScenarioName("Scenario Outline Name First")
 				.setScenarioId("Feature-Name;Scenario-Outline-Name;Examples-Name;10")
-				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":15:16:17")
-				.addScenario().setScenarioKeyword("scenario outline")
-				.setScenarioName("Scenario Outline Name Second")
+				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":15:16:17").addScenario()
+				.setScenarioKeyword("scenario outline").setScenarioName("Scenario Outline Name Second")
 				.setScenarioId("Feature-Name;Scenario-Outline-Name;Examples-Name;20")
-				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":25:26:27")
-				.build();
+				.setScenarioUriStepLines("classpath:features/urisample.feature" + ":25:26:27").build();
 
 		extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest, feature.getElements().get(0));
 		extentTestHeirarchy.createScenarioExtentNode(parentFeatureExtentTest, feature.getElements().get(1));
 
 		List<com.aventstack.extentreports.model.Test> scenarioOutlineTest = parentFeatureExtentTest.getModel()
-				.getNodeContext().getAll();
+				.getChildren();
 		assertEquals("Number of child nodes of Feature Extent Test should be 2", 2, scenarioOutlineTest.size());
 
-		List<com.aventstack.extentreports.model.Test> scenarioFirstTest = scenarioOutlineTest.get(0).getNodeContext()
-				.getAll();
+		List<com.aventstack.extentreports.model.Test> scenarioFirstTest = scenarioOutlineTest.get(0).getChildren();
 		assertEquals("Number of child nodes of First Scenario Outline Test should be 1", 1, scenarioFirstTest.size());
 
-		List<com.aventstack.extentreports.model.Test> scenarioSecondTest = scenarioOutlineTest.get(1).getNodeContext()
-				.getAll();
+		List<com.aventstack.extentreports.model.Test> scenarioSecondTest = scenarioOutlineTest.get(1).getChildren();
 		assertEquals("Number of child nodes of Second Scenario Outline Test should be 1", 1, scenarioSecondTest.size());
 	}
 
@@ -292,23 +260,19 @@ public class ExtentTestHeirarchyTest {
 		String keyword = "Given ";
 		String location = "stepdefs.Stepdefs.step()";
 		String status = "passed";
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario")
-				.setScenarioName("Scenario Name")
-				.addStep().setStepKeyword(keyword).setStepText(stepText)
-				.setStepLocation(location).setStepResult(status)
-				.build();
 
-		ExtentTest stepExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest, feature.getElements().get(0).getSteps().get(0));
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario").setScenarioName("Scenario Name").addStep()
+				.setStepKeyword(keyword).setStepText(stepText).setStepLocation(location).setStepResult(status).build();
 
-		List<com.aventstack.extentreports.model.Test> scenarioChilds = scenarioExtentTest.getModel().getNodeContext()
-				.getAll();
+		ExtentTest stepExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest,
+				feature.getElements().get(0).getSteps().get(0));
+
+		List<com.aventstack.extentreports.model.Test> scenarioChilds = scenarioExtentTest.getModel().getChildren();
 		assertEquals("Number of child nodes of scenario should be 1.", 1, scenarioChilds.size());
 
 		com.aventstack.extentreports.model.Test stepTest = stepExtentTest.getModel();
-		assertEquals("Step keyword is not correct.", keyword.trim(), stepTest.getBehaviorDrivenTypeName());
+		assertEquals("Step keyword is not correct.", "class com.aventstack.extentreports.gherkin.model." + keyword.trim(), stepTest.getBddType().toString());
 		assertEquals("Step text is not correct.", keyword + stepText, stepTest.getName());
 		assertEquals("Step definition method location is not correct.", location, stepTest.getDescription());
 		assertEquals("Step status is not correct.", "PASS", stepTest.getStatus().name());
@@ -320,29 +284,27 @@ public class ExtentTestHeirarchyTest {
 				"Feature Test", "");
 		ExtentTest scenarioExtentTest = featureExtentTest
 				.createNode(com.aventstack.extentreports.gherkin.model.Scenario.class, "Scenario Test", "");
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario")
-				.setScenarioName("Scenario Name")
-				.addStep().setStepKeyword("When ").setStepText("Step text for \u0027when\\u0027 step")
-				.setStepLocation("stepdefs.Stepdefs.stepWhen(java.lang.String)")
-				.setStepResult("failed").setStepErrorMessage("Error Message")
-				.addStep().setStepKeyword("Then ").setStepText("Step text for \u0027then\\u0027 step")
-				.setStepLocation("stepdefs.Stepdefs.stepThen(java.lang.String)").setStepResult("skipped")
-				.build();
+
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario").setScenarioName("Scenario Name").addStep()
+				.setStepKeyword("When ").setStepText("Step text for \u0027when\\u0027 step")
+				.setStepLocation("stepdefs.Stepdefs.stepWhen(java.lang.String)").setStepResult("failed")
+				.setStepErrorMessage("Error Message").addStep().setStepKeyword("Then ")
+				.setStepText("Step text for \u0027then\\u0027 step")
+				.setStepLocation("stepdefs.Stepdefs.stepThen(java.lang.String)").setStepResult("skipped").build();
 
 		Throwable throwable = new Exception("Example Throwable");
 		when(errorMessageProcessor.createThrowableObject(anyString())).thenReturn(throwable);
 
-		ExtentTest stepWhenExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest, feature.getElements().get(0).getSteps().get(0));
+		ExtentTest stepWhenExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest,
+				feature.getElements().get(0).getSteps().get(0));
 		assertEquals("Step status is not correct.", "FAIL", stepWhenExtentTest.getStatus().name());
 
-		ExtentTest stepThenExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest, feature.getElements().get(0).getSteps().get(1));
+		ExtentTest stepThenExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest,
+				feature.getElements().get(0).getSteps().get(1));
 		assertEquals("Step status is not correct.", "SKIP", stepThenExtentTest.getStatus().name());
 
-		List<com.aventstack.extentreports.model.Test> scenarioChilds = scenarioExtentTest.getModel().getNodeContext()
-				.getAll();
+		List<com.aventstack.extentreports.model.Test> scenarioChilds = scenarioExtentTest.getModel().getChildren();
 		assertEquals("Number of child nodes of scenario should be 2.", 2, scenarioChilds.size());
 	}
 
@@ -352,18 +314,15 @@ public class ExtentTestHeirarchyTest {
 				"Feature Test", "");
 		ExtentTest scenarioExtentTest = featureExtentTest
 				.createNode(com.aventstack.extentreports.gherkin.model.Scenario.class, "Scenario Test", "");
-		
-		Feature feature = (new FeatureBuilder())
-				.setUri("classpath:features/urisample.feature").setName("Feature Name")
-				.addScenario().setScenarioKeyword("scenario")
-				.setScenarioName("Scenario Name")
-				.addStep().setStepKeyword("Given ").setStepText("Step text to test")
-				.setStepLocation("stepdefs.Stepdefs.step()").setStepResult("passed")
-				.setStepOutput(Arrays.asList("Information One", "Information Two"))
-				.build();
 
-		ExtentTest stepExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest, feature.getElements().get(0).getSteps().get(0));
-		List<Log> logs = stepExtentTest.getModel().getLogContext().getAll();
+		Feature feature = (new FeatureBuilder()).setUri("classpath:features/urisample.feature").setName("Feature Name")
+				.addScenario().setScenarioKeyword("scenario").setScenarioName("Scenario Name").addStep()
+				.setStepKeyword("Given ").setStepText("Step text to test").setStepLocation("stepdefs.Stepdefs.step()")
+				.setStepResult("passed").setStepOutput(Arrays.asList("Information One", "Information Two")).build();
+
+		ExtentTest stepExtentTest = extentTestHeirarchy.createStepExtentNode(scenarioExtentTest,
+				feature.getElements().get(0).getSteps().get(0));
+		List<Log> logs = stepExtentTest.getModel().getLogs();
 		assertTrue("Step message is not saved in logs.", logs.stream().map(Log::getDetails).collect(Collectors.toList())
 				.containsAll(Arrays.asList("Information One", "Information Two")));
 	}
@@ -376,8 +335,8 @@ public class ExtentTestHeirarchyTest {
 				.createNode(com.aventstack.extentreports.gherkin.model.Scenario.class, "Scenario Test", "");
 
 		String location = "stepdefs.Stepdefs.step()";
-		
-		when(reportProperties.getDisplayAllHooks()).thenReturn(true);
+
+		when(reportProperties.isDisplayAllHooks()).thenReturn(true);
 		Scenario scenario = mock(Scenario.class);
 		Hook hook = mock(Hook.class, RETURNS_DEEP_STUBS);
 		when(scenario.getBefore()).thenReturn(Arrays.asList(hook));
@@ -386,8 +345,7 @@ public class ExtentTestHeirarchyTest {
 
 		List<ExtentTest> hooksExtentTest = extentTestHeirarchy.createBeforeHookExtentNodes(scenarioExtentTest,
 				scenario);
-		List<com.aventstack.extentreports.model.Test> scenarioHooks = scenarioExtentTest.getModel().getNodeContext()
-				.getAll();
+		List<com.aventstack.extentreports.model.Test> scenarioHooks = scenarioExtentTest.getModel().getChildren();
 		assertEquals("Number of hooks of scenario should be 1.", 1, scenarioHooks.size());
 		assertEquals("Hook step definition method is not correct.", location,
 				hooksExtentTest.get(0).getModel().getName());
@@ -401,8 +359,8 @@ public class ExtentTestHeirarchyTest {
 		ExtentTest scenarioExtentTest = featureExtentTest
 				.createNode(com.aventstack.extentreports.gherkin.model.Scenario.class, "Scenario Test", "");
 
-		when(reportProperties.getDisplayAllHooks()).thenReturn(true);
-		
+		when(reportProperties.isDisplayAllHooks()).thenReturn(true);
+
 		Scenario scenario = mock(Scenario.class);
 		Hook hook = mock(Hook.class, RETURNS_DEEP_STUBS);
 		when(scenario.getBefore()).thenReturn(Arrays.asList(hook));
@@ -419,8 +377,8 @@ public class ExtentTestHeirarchyTest {
 		extentTestHeirarchy.createAfterStepHookExtentNodes(scenarioExtentTest, step);
 		extentTestHeirarchy.createAfterHookExtentNodes(scenarioExtentTest, scenario);
 
-		List<com.aventstack.extentreports.model.Test> scenarioHooks = scenarioExtentTest.getModel().getNodeContext()
-				.getAll();
+		List<com.aventstack.extentreports.model.Test> scenarioHooks = scenarioExtentTest.getModel().getChildren();
 		assertEquals("Number of hooks of scenario should be 4.", 4, scenarioHooks.size());
 	}
+
 }